+++ specimen/src/org/labkey/specimen/SpecimenModule.java
@@ -29,10 +29,10 @@ import org.labkey.api.security.roles.RoleManager;
 import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimensPage;
-import org.labkey.specimen.action.ShowSearchAction;
-import org.labkey.specimen.action.SpecimenReportActions;
-import org.labkey.specimen.action.ShowUploadSpecimensAction;
-import org.labkey.specimen.action.ShowGroupMembersAction;
+import org.labkey.specimen.actions.ShowSearchAction;
+import org.labkey.specimen.actions.SpecimenReportActions;
+import org.labkey.specimen.actions.ShowUploadSpecimensAction;
+import org.labkey.specimen.actions.ShowGroupMembersAction;
 import org.labkey.api.specimen.importer.SpecimenImporter;
 import org.labkey.api.specimen.model.SpecimenRequestEvent;
 import org.labkey.api.specimen.model.SpecimenRequestEventType;
@@ -45,9 +45,9 @@ import org.labkey.api.util.emailTemplate.EmailTemplateService;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.HttpView;
 import org.labkey.api.view.WebPartFactory;
-import org.labkey.specimen.action.SpecimenApiController;
-import org.labkey.specimen.action.SpecimenController2;
-import org.labkey.specimen.action.SpecimenController2.OverviewAction;
+import org.labkey.specimen.actions.SpecimenApiController;
+import org.labkey.specimen.actions.SpecimenController2;
+import org.labkey.specimen.actions.SpecimenController2.OverviewAction;
 import org.labkey.specimen.importer.DefaultSpecimenImportStrategyFactory;
 import org.labkey.specimen.importer.SpecimenSchemaImporter;
 import org.labkey.specimen.importer.SpecimenSettingsImporter;
