+++ specimen/src/org/labkey/specimen/actions/ShowSearchAction.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.labkey.specimen.action;
+package org.labkey.specimen.actions;
 
 import org.labkey.api.action.FormViewAction;
 import org.labkey.api.data.CompareType;
