+++ specimen/src/org/labkey/specimen/actions/SpecimenApiController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.labkey.specimen.action;
+package org.labkey.specimen.actions;
 
 import org.apache.commons.lang3.StringUtils;
 import org.json.JSONObject;
