+++ specimen/src/org/labkey/specimen/actions/SpecimenReportActions.java
@@ -1,4 +1,4 @@
-package org.labkey.specimen.action;
+package org.labkey.specimen.actions;
 
 import org.labkey.api.security.RequiresPermission;
 import org.labkey.api.security.permissions.ReadPermission;
