+++ specimen/src/org/labkey/specimen/report/participant/ParticipantTypeReportFactory.java
@@ -24,7 +24,7 @@ import org.labkey.api.specimen.model.SpecimenTypeSummary;
 import org.labkey.specimen.report.SpecimenVisitReport;
 import org.labkey.specimen.report.SpecimenVisitReportAction;
 import org.labkey.specimen.report.SpecimenVisitReportParameters;
-import org.labkey.specimen.action.SpecimenReportActions;
+import org.labkey.specimen.actions.SpecimenReportActions;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.Visit;
 import org.labkey.api.util.HtmlString;
