+++ specimen/src/org/labkey/specimen/report/request/RequestReportFactory.java
@@ -19,7 +19,7 @@ import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.specimen.report.SpecimenVisitReport;
 import org.labkey.specimen.report.SpecimenVisitReportAction;
-import org.labkey.specimen.action.SpecimenReportActions;
+import org.labkey.specimen.actions.SpecimenReportActions;
 import org.labkey.api.study.Location;
 import org.labkey.api.study.Visit;
