+++ specimen/src/org/labkey/specimen/report/specimentype/TypeParticipantReportFactory.java
@@ -22,7 +22,7 @@ import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.specimen.report.SpecimenTypeVisitReport;
 import org.labkey.specimen.report.SpecimenVisitReport;
 import org.labkey.specimen.report.SpecimenVisitReportAction;
-import org.labkey.specimen.action.SpecimenReportActions;
+import org.labkey.specimen.actions.SpecimenReportActions;
 import org.labkey.api.study.Cohort;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
