+++ specimen/src/org/labkey/specimen/report/specimentype/TypeSummaryReportFactory.java
@@ -20,7 +20,7 @@ import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.specimen.report.SpecimenTypeVisitReport;
 import org.labkey.specimen.report.SpecimenVisitReport;
 import org.labkey.specimen.report.SpecimenVisitReportAction;
-import org.labkey.specimen.action.SpecimenReportActions;
+import org.labkey.specimen.actions.SpecimenReportActions;
 import org.labkey.api.study.Visit;
 
 import java.util.Collections;
