+++ specimen/src/org/labkey/specimen/view/autoReportList.jsp
@@ -20,7 +20,7 @@
 <%@ page import="org.labkey.api.data.Container" %>
 <%@ page import="org.labkey.api.query.CustomView" %>
 <%@ page import="org.labkey.api.security.User" %>
-<%@ page import="org.labkey.specimen.action.ReportConfigurationBean" %>
+<%@ page import="org.labkey.specimen.actions.ReportConfigurationBean" %>
 <%@ page import="org.labkey.specimen.report.SpecimenVisitReportParameters" %>
 <%@ page import="org.labkey.api.study.Cohort" %>
 <%@ page import="org.labkey.api.study.CohortFilter" %>
