+++ specimen/src/org/labkey/specimen/view/manageSpecimenWebPart.jsp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 %>
-<%@ page import="org.labkey.specimen.action.SpecimenController2.SpecimenWebPartForm" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController2.SpecimenWebPartForm" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
