+++ specimen/src/org/labkey/specimen/view/uploadSimpleSpecimens.jsp
@@ -21,8 +21,8 @@
 <%@ page import="org.labkey.api.study.TimepointType" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
-<%@ page import="org.labkey.specimen.action.ShowUploadSpecimensAction" %>
-<%@ page import="org.labkey.specimen.action.SpecimenController2.GetSpecimenExcelAction" %>
+<%@ page import="org.labkey.specimen.actions.ShowUploadSpecimensAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController2.GetSpecimenExcelAction" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%
